services/auth-python/requirements.txt
fastapi==0.115.2
uvicorn==0.30.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
pymysql==1.1.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pyotp==2.9.0
python-multipart==0.0.9
cryptography==43.0.1


